Clean up according to PR checklist
Check the keymaps/nuess0r how you can make better use of your Model M
including Windows, multimedia keys etc.
