+++ keyboards/ibm/model_m/ctrl_m/ctrl_m.c
@@ -14,22 +14,3 @@
  * along with this program.  If not, see <http://www./gnu.org/licenses/>.
  */
 #include "ctrl_m.h"
-
-void keyboard_pre_init_kb(void) {
-  /* Setting status LEDs pins to output and +5V (off) */
-  setPinOutput(LED_NUM_LOCK_PIN);
-  setPinOutput(LED_CAPS_LOCK_PIN);
-  setPinOutput(LED_SCROLL_LOCK_PIN);
-  writePinHigh(LED_NUM_LOCK_PIN);
-  writePinHigh(LED_CAPS_LOCK_PIN);
-  writePinHigh(LED_SCROLL_LOCK_PIN);
-}
-
-bool led_update_kb(led_t led_state) {
-    if(led_update_user(led_state)) {
-        writePin(LED_NUM_LOCK_PIN, !led_state.num_lock);
-        writePin(LED_CAPS_LOCK_PIN, !led_state.caps_lock);
-        writePin(LED_SCROLL_LOCK_PIN, !led_state.scroll_lock);
-    }
-    return true;
-}
