+++ keyboards/ibm/model_m/ctrl_m/keymaps/nuess0r/config.h
@@ -17,5 +17,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #pragma once
 
-/* The Model M as a membrane based keyboard without diodes has ghost keys */
-#define MATRIX_HAS_GHOST
+/* As we replace the caps lock key with the win key, enable by default the
+ * Caps Word feature by double tapping the shift key */
+#define DOUBLE_TAP_SHIFT_TURNS_ON_CAPS_WORD
