+++ keyboards/ibm/model_m/ctrl_m/keymaps/nuess0r/rules.mk
@@ -1,0 +1,1 @@
+VIA_ENABLE = yes
