+++ keyboards/ibm/model_m/ctrl_m/readme.md
@@ -11,8 +11,20 @@ Example how to compile for this keyboard (after setting up your build environmen
 
     make ibm/model_m/ctrl_m:default
 
+Flashing example for this keyboard:
+
+    make %KEYBOARD%:default:flash
+
 Or if you use the QMK tool:
 
     qmk compile -kb ibm/model_m/ctrl_m -km default
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
+
+## Bootloader
+
+Enter the bootloader in 3 ways:
+
+* **Bootmagic reset**: Hold down the Escape key and plug in the keyboard
+* **Physical reset button**: Briefly press the button on the back of the PCB and plug in the keyboard
+* **Keycode in layout**: Press the key mapped to `QK_BOOT` if it is available
